Create problem13.py
Accept a positive integer 
�
n that represents the number of students in the class. 
�
n blocks of input follow. Each block is made up of six lines and contains the details of one student in the class. Create a dictionary corresponding to each student. All keys should be strings. The type of the value corresponding to a key and the order in which the inputs should be accepted are shown in the table given below.